Separated the auto folder into a sensor and PID foler
